+++ App.js
@@ -1,30 +1,11 @@
 // @flow
+/* eslint-disable */
 
 import React from 'react';
-import { StyleSheet, Button, View } from 'react-native';
+import Home from './src/config/router';
 
-const onPress = () => true;
+const App = () => <Home />;
 
-export default () => (
-  <View style={styles.container}>
-    <Button
-      onPress={onPress}
-      title="Import Funds"
-      accessibilityLabel="Import funds from another Ethereum wallet"
-    />
-    <Button
-      onPress={onPress}
-      title="Create New Wallet"
-      accessibilityLabel="Create new Ethereum wallet"
-    />
-  </View>
-);
+export default App;
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+
