+++ package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "expo": "^25.0.0",
     "react": "16.2.0",
-    "react-native": "0.52.2"
+    "react-native": "0.52.2",
+    "react-navigation": "^1.0.0-beta.27"
   }
 }
